+++ src/Stash/Driver/Predis.php
@@ -33,8 +33,9 @@ class Predis implements Driver
     /**
      * Init with settings
      */
-    public function __construct(array $settings)
-    {
+    public function __construct(
+        array $settings
+    ) {
         $this->generatePrefix(
             Coercion::toStringOrNull($settings['prefix'] ?? null)
         );
@@ -120,8 +121,9 @@ class Predis implements Driver
     /**
      * Clear all values from store
      */
-    public function clearAll(string $namespace): bool
-    {
+    public function clearAll(
+        string $namespace
+    ): bool {
         $key = $this->createNestedKey($namespace, null)[1];
 
         if (!$this->client->incr($key)) {
@@ -170,12 +172,30 @@ class Predis implements Driver
     }
 
 
+    /**
+     * Count items
+     */
+    public function count(
+        string $namespace,
+    ): int {
+        return count($this->getKeys($namespace));
+    }
+
+    /**
+     * Get key
+     */
+    public function getKeys(string $namespace): array
+    {
+        return $this->client->keys($this->prefix . ':*');
+    }
+
 
     /**
      * Get cached path index
      */
-    protected function getPathIndex(string $pathKey): int
-    {
+    protected function getPathIndex(
+        string $pathKey
+    ): int {
         return (int)$this->client->get($pathKey);
     }