+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v0.2.0 (2023-11-06)
+* Added Dovetail config implementation
+* Converted PileUpPolicy to enum
+* Added getKeys() and count() to Store and Driver
+* Added default prefix to Context
+* Renamed get() to load()
+* Refactored package file structure
+* Updated Genesis dependency
+
 ## v0.1.2 (2023-09-26)
 * Converted phpstan doc comments to generic
 * Migrated to use effigy in CI workflow